+++ src/main/java/com/example/flower_shop/controller/FlowerController.java
@@ -30,20 +30,20 @@ public class FlowerController {
     }
    
     @GetMapping("/{id}")
-    public Flower getFlowerById(@PathVariable Long id) {
+    public Flower getFlowerById(@PathVariable Integer id) {
         return flowerRepository.findById(id).orElse(null);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT) // kthe 204 per delete sepse nuk ka body
-    public void deleteFlower(@PathVariable Long id) {
+    public void deleteFlower(@PathVariable Integer id) {
         if(flowerRepository.existsById(id)){
             flowerRepository.deleteById(id);
         } 
     }
 
     @PutMapping("/{id}")
-    public Flower updateFlower(@PathVariable Long id, @RequestBody Flower flowerDetails) {
+    public Flower updateFlower(@PathVariable Integer id, @RequestBody Flower flowerDetails) {
         Flower flower = flowerRepository.findById(id).orElse(null);
         if (flower != null) {
             flower.setEmertimi(flowerDetails.getEmertimi());
